requirements.txt
fastapi[standard]
sqlalchemy
uvicorn
pydantic
passlib
bcrypt
python-jose